requirements.txt
Flask==2.3.0
flask-sqlalchemy==3.1.1
flask_session